tools: add deflated size of the zipfile to the updates.xml
The size returned is the sum of the sizes of all the files in the
zip file.

NOTE: This is not exactly the same as disk usage after decompression.
The size returned here does not take into account the real disk usage of
the files (dependent on the fs block sizes) or disk space taken by
directories.

Change-Id: I8c52aed054c6fe9cf7ec9c3add24c0395cbb5c79
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>