Ensure only new updates are installed
The previously get_uninstalled_updates would look only at updates
which had been installed by the system updater and would then add
updates that didn't need to be applied because the current system
was installed from a later image than the earlier updates.

This patch adds a check for said invalid updates as well as creates
a folder for the zypper rpm cache of the update when needed.

Signed-off-by: William Douglas <william.douglas@intel.com>